Few minor changes: 1. add missing icon 2. remove some data from elements to make it more clear 3. small changes in namings
