+++ dotbimGH/Properties/Resources.resx
@@ -52,4 +52,7 @@
     <data name="CreateSimpleElementsSet" type="System.Resources.ResXFileRef, System.Windows.Forms">
         <value>..\Resources\CreateSimpleElementsSet.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
     </data>
+    <data name="BakeFile" type="System.Resources.ResXFileRef, System.Windows.Forms">
+        <value>..\Resources\BakeFile.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+    </data>
 </root>
