+++ dotbimGH/dotbimGH.csproj
@@ -34,7 +34,7 @@
   <ItemGroup>
     <Compile Include="BimElement.cs" />
     <Compile Include="BimElementSet.cs" />
-    <Compile Include="Components\BakeElement.cs" />
+    <Compile Include="Components\BakeFile.cs" />
     <Compile Include="Components\CreateElementGh.cs" />
     <Compile Include="Components\CreateElementSetGh.cs" />
     <Compile Include="Components\CreateFileGh.cs" />
@@ -63,6 +63,7 @@
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
+    <Content Include="Resources\BakeFile.png" />
     <Content Include="Resources\CreateElement.png" />
     <Content Include="Resources\CreateElementsSet.png" />
     <Content Include="Resources\CreateFile.png" />
